this it it

+++ backend/api/taskController.py
@@ -6,7 +6,9 @@ from flask import request
 from api.resource import ApiResource
 from domain.task import Task
 from infra.taskRepository import TaskRepository
+from infra.userRepository import UserRepository
 
+user_repository = UserRepository()
 task_repository = TaskRepository()
 
 
@@ -28,6 +30,8 @@ class TaskController(ApiResource):
         return "/task"
 
     def post(self):
+        if(not user_repository.is_valide_token(request.headers.get("authorization").replace("Bearer ", ""))):
+            return {"error": "Token invalide"}, 400
         data = request.get_json()
         if not is_valid_task_format(data):
             return {'invalid_format': True}
@@ -35,6 +39,8 @@ class TaskController(ApiResource):
         return data
 
     def get(self):
+        if(not user_repository.is_valide_token(request.headers.get("authorization").replace("Bearer ", ""))):
+            return {"error": "Token invalide"}, 400
         tasks = task_repository.findAll()
         tasks_json = json.loads(json_util.dumps(tasks))
         return tasks_json
@@ -46,10 +52,14 @@ class TaskIdController(ApiResource):
         return "/task/<string:task_id>"
 
     def delete(self, task_id):
+        if(not user_repository.is_valide_token(request.headers.get("authorization").replace("Bearer ", ""))):
+            return {"error": "Token invalide"}, 400
         task_repository.delete(task_id)
         return f"Task {task_id} deleted"
 
     def put(self, task_id):
+        if(not user_repository.is_valide_token(request.headers.get("authorization").replace("Bearer ", ""))):
+            return {"error": "Token invalide"}, 400
         data = request.get_json()
         if is_valid_task_format(data):
             task_repository.put(data_to_task(data), task_id)
